server/v1/pyproject.toml
[tool.poetry]
name = "goteamapi"
version = "0.1.0"
description = "GoTeam! application's backend."
authors = ["Alican Demirtas <demirtasdev@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "^3.3.4"
astroid = "^2.5.6"
attrs = "^21.2.0"
bcrypt = "^3.2.0"
cffi = "^1.14.5"
colorama = "^0.4.4"
dj-database-url = "^0.5.0"
Django = "^3.2.2"
django-cors-headers = "^3.7.0"
django-filter = "^2.4.0"
django-pgviews = "^0.5.7"
djangorestframework = "^3.12.4"
docopt = "^0.6.2"
gunicorn = "^20.1.0"
iniconfig = "^1.1.1"
isort = "^5.8.0"
lazy-object-proxy = "^1.6.0"
Markdown = "^3.3.4"
mccabe = "^0.6.1"
Pillow = "^8.2.0"
pluggy = "^0.13.1"
psycopg2-binary = "^2.8.6"
py = "^1.10.0"
pycparser = "^2.20"
pyparsing = "^2.4.7"
python-dotenv = "^0.17.1"
python-status = "^1.0.1"
pytz = "^2021.1"
rope = "^0.19.0"
six = "^1.16.0"
sqlparse = "^0.4.1"
toml = "^0.10.2"
watchdog = "^2.1.1"
whitenoise = "^5.2.0"
wrapt = "^1.12.1"
shortuuid = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
